requirements.txt
# Install triton nightly for CUDA compatibility
--find-links https://download.pytorch.org/whl/nightly/cu124
triton-nightly

# Core dependencies
openai
jinja2
python-dotenv
gradio

# Utility Dependencies
pathlib2>=2.3.0; python_version < '3.4'

# Development Dependencies (optional)
# pytest>=7.0.0
# flake8>=6.0.0
# black>=23.0.0 